Extracting obvious reporting code to a Reporter class.
This is a first cleanup pass, more will follow. This should allow to
switch between Reporters, to have different kind of output (including
null output).

This first pass extract obvious reporting code to a single class,
without much changes to the current structure of the code. It
already moves all access to tqdm in a single place. More reporting
done via logging in various methods still needs to be moved.

Bug: T212783
Change-Id: Iabdf4904e9855544557158f92be7ba7b23070c50