transport.clustershell: handle str when reporting commands
When the callback methods (ev_*) are called by clustershell, the
worker passed has a property 'command' that is a string, all the other
methods use 'self.commands[some_index]' and that's a Command object.
We were only taking into account the Command objects but forgot to
handle the 'str' passed by clustershell.

Bug: T275210
Change-Id: I0cf430ec301f0151712beeec7f7bd15dc85fc7d7
Signed-off-by: David Caro <david@dcaro.es>